Редусер переразбит иначе
