+++ package.json
@@ -23,7 +23,7 @@
     "react-test-renderer": "16.11.0",
     "webpack": "4.41.2",
     "webpack-cli": "3.3.9",
-    "webpack-dev-server": "3.8.2"
+    "webpack-dev-server": "^3.8.2"
   },
   "engines": {
     "node": "10"
