+++ src/components/app/app.jsx
@@ -1,23 +1,33 @@
-import React from 'react';
+import React, {PureComponent} from 'react';
 import PropTypes from 'prop-types';
 import {connect} from 'react-redux';
-import {ActionCreator as ActionCreatorData} from '../../reducer/data/action-creator';
-import {Operation as UserOperation} from '../../reducer/user/user.js';
-import {getAllOffers, getIsOffersLoading, getCurrentCity, getOffers, getCitiesState} from '../../reducer/data/selectors';
+import * as ActionCreatorData from '../../reducer/data/action-creator.js';
+import * as DataOperation from '../../reducer/data/operations.js';
+import * as UserOperation from '../../reducer/user/operations.js';
+import {getAllOffers, getIsOffersLoading, getCurrentCity, getOffers, getCitiesState, getReviews} from '../../reducer/data/selectors';
 import {getIsAuthorizationRequired, getUserInfo} from '../../reducer/user/selectors';
-import {SignIn} from '../sign-in/sign-in.jsx';
+// import {SignIn} from '../sign-in/sign-in.jsx';
 import {PageScreen} from '../page-screen/page-screen.jsx';
 
-const App = (props) => {
-  const {isAuthorizationRequired, authenticateUser} = props;
-  if (isAuthorizationRequired) {
-    return <SignIn isAuthorizationRequired={isAuthorizationRequired} authenticateUser={authenticateUser}/>;
+class App extends PureComponent {
+
+  componentDidMount() {
+    this.props.loadAllOffers();
   }
-  if (props.isOffersLoading) {
-    return null;
+
+  render() {
+    if (this.props.isOffersLoading) {
+      return null;
+    }
+    return <PageScreen {...this.props}/>;
   }
-  return <PageScreen {...props}/>;
-};
+  // const {isAuthorizationRequired, authenticateUser} = props;
+  // if (isAuthorizationRequired) {
+  //   return <SignIn isAuthorizationRequired={isAuthorizationRequired} authenticateUser={authenticateUser}/>;
+  // }
+
+}
+
 
 const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
   allOffers: getAllOffers(state),
@@ -27,9 +37,13 @@ const mapStateToProps = (state, ownProps) => Object.assign({}, ownProps, {
   cities: getCitiesState(state),
   isAuthorizationRequired: getIsAuthorizationRequired(state),
   userInfo: getUserInfo(state),
+  reviews: getReviews(state),
+  // isReviewsLoading: getIsReviewsLoading(state),
 });
 
 const mapDispatchToProps = (dispatch) => ({
+  loadAllOffers: () => dispatch(DataOperation.loadAllOffers()),
+
   authenticateUser: (email, password) => {
     // dispatch(ActionCreatorUser.authenticateUser(email, password));
     dispatch(UserOperation.authenticateUser(email, password));
@@ -37,10 +51,14 @@ const mapDispatchToProps = (dispatch) => ({
   onCityClick: (currentCity, allOffers) => {
     dispatch(ActionCreatorData.changeCity(currentCity));
     dispatch(ActionCreatorData.getOffersList(currentCity, allOffers));
+  },
+  loadOfferReviews: (id) => {
+    dispatch(DataOperation.loadOfferReviews(id));
   }
 });
 
 App.propTypes = {
+  loadAllOffers: PropTypes.func,
   isAuthorizationRequired: PropTypes.bool,
   authenticateUser: PropTypes.func,
   isOffersLoading: PropTypes.bool,
