+++ src/index.js
@@ -6,7 +6,7 @@ import thunk from 'redux-thunk';
 import {compose} from 'recompose';
 import App from './components/app/app.jsx';
 import createAPI from './api';
-import {Operation} from './reducer/data/data.js';
+// import * as Operation from './reducer/data/operations.js';
 import reducer from './reducer';
 
 const init = () => {
@@ -19,7 +19,7 @@ const init = () => {
       )
   );
 
-  store.dispatch(Operation.loadAllOffers());
+  // store.dispatch(Operation.loadAllOffers());
 
   ReactDOM.render(
       <Provider store={store}>
