+++ src/reducer/data/action-type.js
@@ -2,6 +2,8 @@ export const ActionType = {
   LOAD_ALL_OFFERS: `LOAD_ALL_OFFERS`,
   CHANGE_LOADING_STATE: `CHANGE_LOADING_STATE`,
   CHANGE_CITY: `CHANGE_CITY`,
-  GET_OFFERS_LIST: `GET_OFFERS_LIST`,
   GET_CITIES_LIST: `GET_CITIES_LIST`,
+  GET_OFFERS_LIST: `GET_OFFERS_LIST`,
+  LOAD_OFFER_REVIEWS: `GET_OFFER_REVIEWS`,
+  // CHANGE_REVIEWS_LOADING_STATE: `CHANGE_REVIEWS_LOADING_STATE`,
 };
