+++ src/reducer/data/initial-state.js
@@ -1,0 +1,8 @@
+export default {
+  allOffers: [],
+  isOffersLoading: true,
+  currentCity: null,
+  offers: [],
+  cities: [],
+  reviews: null,
+};
