+++ src/reducer/data/prepare-data.js
@@ -1,0 +1,41 @@
+
+export const prepareOffers = (allOffers) => {
+  const preparedData = allOffers.map((offer) => {
+    return {
+      id: offer.id,
+      city: {
+        name: offer.city.name,
+        location: {
+          latitude: offer.city.location.latitude,
+          longitude: offer.city.location.longitude,
+          zoom: offer.city.location.zoom,
+        }
+      },
+      previewImage: offer.preview_image,
+      images: offer.images,
+      title: offer.title,
+      isFavorite: offer.is_favorite,
+      isPremium: offer.is_premium,
+      rating: offer.rating,
+      type: offer.type,
+      bedrooms: offer.bedrooms,
+      maxAdults: offer.max_adults,
+      price: offer.price,
+      goods: offer.goods,
+      host: {
+        id: offer.host.id,
+        name: offer.host.name,
+        isPro: offer.host.is_pro,
+        avatarUrl: offer.host.avatar_url,
+      },
+      description: offer.description,
+      location: {
+        latitude: offer.location.latitude,
+        longitude: offer.location.longitude,
+        zoom: offer.location.zoom,
+      },
+    };
+  });
+
+  return preparedData;
+};
