+++ src/reducer/index.js
@@ -1,6 +1,6 @@
 import {combineReducers} from 'redux';
-import {reducer as data} from './data/data';
-import {reducer as user} from './user/user';
+import {reducer as data} from './data/reducer';
+import {reducer as user} from './user/reducer';
 import NameSpace from './name-spaces';
 
 export default combineReducers({
