+++ src/reducer/user/initial-state.js
@@ -1,0 +1,6 @@
+export default {
+  isAuthorizationRequired: false,
+  email: ``,
+  password: ``,
+  userInfo: null,
+};
